+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 news_categories Changelog
 =========================
 
+Version 2.8.0 stable (2017-01-27)
+---------------------------------
+
+### Improved
+- Allow to hide category in the news list module
+- Allow to hide category in the news reader module
+- Allow to exclude the news of category in the related list module
+
+
 Version 2.7.5 stable (2016-10-20)
 ---------------------------------